requirements.txt
langchain
python-dotenv
openai
langchain-community
streamlit
sentence-transformers
faiss-cpu
google-search-results
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14